chore(api): migrate api entrypoint to TypeScript
