+++ api/index.js
@@ -1,5 +1,1 @@
-const { createApp } = require("../backend/server.cjs");
-
-const app = createApp();
-
-module.exports = (req, res) => app.handleRequest(req, res);
+module.exports = require('../.tsbuild/api/index.cjs');
